+++ src/scripts/common.js
@@ -43,3 +43,17 @@ export const sleep = function(amount) {
     setTimeout(resolve, amount)
   })
 }
+
+export function toMicroDenom(denom) {
+  if (denom === "ATOM") {
+    return "uatom"
+  }
+  return denom.toLowerCase()
+}
+
+export function fromMicroDenom(denom) {
+  if (denom.toLowerCase() === "uatom") {
+    return "ATOM"
+  }
+  return denom.toUpperCase()
+}
